inclusão do serviço cidades

+++ src/main/java/br/uema/engcomp/oficina/model/Cliente.java
@@ -35,7 +35,7 @@ public class Cliente implements Serializable {
     @NotBlank(message = "O nome é obrigatório!")
     private String nome;
     private String endereco;
-    private int cidades_id;
+    private Long cidades_id;
     private String telefone;
     private String email;
     private String dt_nasc;
